% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend_functions.R
\name{aggregate_where}
\alias{aggregate_where}
\title{Conditional aggregation}
\usage{
aggregate_where(fun, var, cond)
}
\arguments{
\item{fun}{The aggregation function to run}

\item{var}{The variable to run the function on}

\item{cond}{The condition to filter on}
}
\description{
Aggregate a column where some condition is true - meant to be used inside tidy functions. 
This is not even a convenience function as it can be done simpler in native tidyverse (se examples).
The function exists in order to highlight a programming pattern.
}
\examples{
\dontrun{
mutate(msft, fincol = aggregate_where(mean, adj_close, change=="UP") )

# Equivalent to
mutate(msft, fincol = mean(adj_close[change=="UP"] ))
}

}
\keyword{aggregation}
