% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{match_number}
\alias{match_number}
\title{Match number
This is an internal function that returns an array of match-numbers, in the case that 
there are several matches in one "string".}
\usage{
match_number(matchlist)
}
\arguments{
\item{matchlist}{Match ranges}
}
\description{
Match number
This is an internal function that returns an array of match-numbers, in the case that 
there are several matches in one "string".
}
