% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_functions.R
\name{match_partition}
\alias{match_partition}
\title{Match rows per group}
\usage{
match_partition(df, definitions, rx, match_name = NULL)
}
\arguments{
\item{df}{Sorted dataframe to filter (this will often be a partition from a group_by statement)}

\item{definitions}{Column containing the definition of rows}

\item{rx}{Simple regex-like statement to filter for - quoted.}

\item{match_name}{Optional column name for match-number, defaults to NULL.}
}
\description{
This function accepts a dataframe, regex-pattern and column name for definitions, and return matching rows
The function is meant to be called from match_rows, due to quosures it wont work standalone.
It is, however, fairly simple to call this from another wrapper function as long as it takes care of quoting.
}
