% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_functions.R
\name{match_partition_raw}
\alias{match_partition_raw}
\title{Regex row matcher per partition}
\usage{
match_partition_raw(df, definitions, rx, match_name = NULL)
}
\arguments{
\item{df}{Sorted and grouped dataframe to filter}

\item{match_name}{Optional column name for match-number, defaults to NULL.}

\item{ptn}{Perl-regex pattern to look for, based on the definitions created}

\item{defs}{Column containing the definition of rows}
}
\description{
This function is meant to be called from match_rows_raw which handles partitions.
The function accepts a dataframe, regex-pattern and column name for definitions, and return matching rows
}
