% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend_functions.R
\name{match_rows_raw}
\alias{match_rows_raw}
\title{Regex row matcher}
\usage{
match_rows_raw(df, definitions, rx, match_name = NULL)
}
\arguments{
\item{df}{Sorted and grouped dataframe to filter}

\item{match_name}{Optional column name for match-number, defaults to NULL.}

\item{ptn}{Perl-regex pattern to look for, based on the definitions created}

\item{defs}{Column containing the definition of rows}
}
\description{
This function accepts a dataframe, regex-pattern and column name for definitions, and return matching rows
}
\examples{
\dontrun{
match_rows_raw(msft, change2, "([D]{4,})", mnum)
}
}
\keyword{match_recognize}
