% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msft}
\alias{msft}
\title{Stock data for Microsoft}
\description{
A subset of the `stocks` dataset, pre-sorted and with example definition columns for ease of use in examples.
The new variables are:
}
\details{
\itemize{
  \item change. Character variable cointaining "UP" if `adj_close` is higher than previous day, else "DOWN".
  \item change2. Same as `change`, but containing only "U" and "D" for use with `match_rows_raw` function.
}
}
\references{
\url{https://finance.yahoo.com/quotes/AMZN,AAPL,MSFT}
}
\author{
Radbrt \email{dev_null@cerx.co}
}
\keyword{datasets}
