% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ranges_from_starts_lengths}
\alias{ranges_from_starts_lengths}
\title{Match ranges
DEPRECATED
This is a convenience function that recieves two integers (a, b), and returns the range a:b-1}
\usage{
ranges_from_starts_lengths(start, ends)
}
\arguments{
\item{start}{The start of the range}

\item{ends}{The end of the range, endexclusive}
}
\description{
Match ranges
DEPRECATED
This is a convenience function that recieves two integers (a, b), and returns the range a:b-1
}
