% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{regex_row_matcher}
\alias{regex_row_matcher}
\title{Regex row matcher}
\usage{
regex_row_matcher(df, ptn, defs)
}
\arguments{
\item{df}{Sorted and grouped dataframe to filter}

\item{ptn}{Perl-regex pattern to look for, based on the definitions created}

\item{defs}{Column containing the definition of rows}
}
\description{
This function accepts a dataframe, regex-pattern and column name for definitions, and return matching rows
}
\examples{
\dontrun{
regex_row_matcher(stocks, "([D]{4,})", ds)
}
}
\keyword{match_recognize}
