% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{row_ranges}
\alias{row_ranges}
\title{Ranges from pattern
returns row ranges e.g. c(1:4,13:17) based on string of definitions and regex-pattern}
\usage{
row_ranges(defstring, ptn)
}
\arguments{
\item{defstring}{The collapsed column with definitions - one character per row}

\item{ptn}{The regex-pattern to find}
}
\description{
Ranges from pattern
returns row ranges e.g. c(1:4,13:17) based on string of definitions and regex-pattern
}
