% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{subset_from_ranges}
\alias{subset_from_ranges}
\title{Subset a dataframe based on list of possibly overlapping ranges of rows}
\usage{
subset_from_ranges(df, ranges)
}
\arguments{
\item{df}{Data frame to subset}

\item{ranges}{List of possibly overlapping ranges of rows to select}
}
\description{
Subset a dataframe based on list of possibly overlapping ranges of rows
}
