% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_functions.R
\name{match_partition_raw}
\alias{match_partition_raw}
\title{Regex row matcher per partition}
\usage{
match_partition_raw(df, definitions, rx, match_name = NULL,
  keep_all_rows = FALSE)
}
\arguments{
\item{df}{Sorted and grouped dataframe to filter}

\item{definitions}{Column containing the definition of rows}

\item{rx}{Perl-regex pattern to look for, based on the definitions created}

\item{match_name}{Optional column name for match-number, defaults to NULL.}

\item{keep_all_rows}{Boolean allows you to return all rows, uncluding nonmatching rows.}
}
\description{
This function is meant to be called from match_rows_raw which handles partitions.
The function accepts a dataframe, regex-pattern and column name for definitions, and return matching rows
}
