% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{match_rows_raw}
\alias{match_rows_raw}
\title{Regex row matcher}
\usage{
match_rows_raw(df, defs, ptn)
}
\arguments{
\item{df}{Sorted and grouped dataframe to filter}

\item{defs}{Column containing the definition of rows}

\item{ptn}{Perl-regex pattern to look for, based on the definitions created}
}
\description{
This function accepts a dataframe, regex-pattern and column name for definitions, and return matching rows
}
\examples{
\dontrun{
match_rows_raw(stocks, ds, "([D]{4,})")
}
}
\keyword{match_recognize}
