% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\references{
\url{yahoo.com}
}
\author{
Radbrt \email{dev_null@cerx.co}
}
\keyword{data}
\keyword{stock}
